+++ src/WelcomeBot.js
@@ -111,6 +111,7 @@ class WelcomeBot extends Client {
         this.botServerId = "836854115526770708";
         this.newsChannelId = "847459283876577360";
         this.loggingChannelId = "855331801635749888";
+        this.suggestionLogsChannelId = "862126837110800414";
         this.ownerIDs = [
             "815204465937481749" /*PuneetGopinath#0001*/,
             "693754859014324295" /*abhijoshi2k#6842*/,