+++ src/commands/setup/enable.js
@@ -31,6 +31,7 @@ module.exports = {
                 message.react("👍");
                 break;
             default:
+            if (!args.length) {
                 return message.channel.send(
                     `Welcome logs are ${
                         guildDB.enableWelcome ? "enabled" : "disabled"
@@ -38,6 +39,12 @@ module.exports = {
                         guildDB.enableGoodbye ? "enabled" : "disabled"
                     }`
                 );
+                } else {
+                    message.reply(
+                        t("cmds:channel.invalidArgs") +
+                            `${guildDB.prefix}help enable`
+                    );
+                }
                 break;
         }
         return;